Extract tests into relevant class
