+++ src/main/java/com/github/codeidoscope/InputValidator.java
@@ -1,6 +1,22 @@
 package com.github.codeidoscope;
 
-public class InputValidator {
+class InputValidator {
+
+    void checkArgumentLength(String[] arguments){
+        if (arguments.length != 2) {
+            System.err.println("Missing port number");
+            System.exit(1);
+        }
+    }
+
+    void checkPortNumberIsValid(String portNumberInput) {
+        try {
+            Integer.parseInt(portNumberInput);
+        } catch (NumberFormatException e) {
+            System.err.println("Incorrect port number. Please try again");
+            System.exit(1);
+        }
+    }
 
     String getInputAndPrepareOutput(String inputLine) {
